+++ public/src/jsPopUps/popUpData.js
@@ -154,7 +154,7 @@ export const popUpData = {
      </table>
      <button data-send id="send">Verzenden</button>
       <div data-markerContainer class="markerContainer center noDisplay">
-        <div class="arrowArm loader spin"></div>
+        <div class="windSockArm loader spin"></div>
         <div class="marker loader"></div>
       </div>`,
 	},